requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django

django-environ==0.11.2
django-pagination==1.0.7
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
